+++ camera.h
@@ -18,8 +18,8 @@ public:
     inline Vec3 dx () {return _dx;}
     inline Vec3 dy () {return _dy;}
     inline Vec3 environment () {return _environment;}
-    inline float focus () {return _focus;}
-    Ray emitRay (float x, float y);
+    inline double focus () {return _focus;}
+    Ray emitRay (double x, double y);
 protected:
     void analyseContent (const std::string& content);
     void init ();
@@ -31,7 +31,7 @@ private:
     int _dpi;
     Vec3 _dx, _dy;//镜头x,y正方向的单位向量
     Vec3 _center, _normal;//中心，单位法向量
-    float _focus;//感光点到image的距离
+    double _focus;//感光点到image的距离
     Color _environment;
 };
 std::ostream& operator<< (std::ostream& os, const Camera& camera);
