+++ object.h
@@ -27,11 +27,11 @@ protected:
 class Sphere: public Object
 {
 public:
-    Sphere (const Vec3& center, float radius, Material* material = nullptr, Condutor* condutor = nullptr, bool need = false);
+    Sphere (const Vec3& center, double radius, Material* material = nullptr, Condutor* condutor = nullptr, bool need = false);
     Sphere (std::stringstream& content, Condutor* condutor = nullptr);
     virtual ~Sphere ();
     Vec3 center () const {return _center;}
-    float radius () const {return _radius;}
+    double radius () const {return _radius;}
     virtual Collide collide (const Ray &ray) const;
     virtual void display (std::ostream& os) const;
     virtual Color color (const Vec3 &v = Vec3 ()) const;
@@ -42,9 +42,9 @@ protected:
     void preHandle ();
 private:
     Vec3 _center;
-    float _radius;
+    double _radius;
 
-    float _r2;
+    double _r2;
     bool _needBoudingBox;
 };
 
@@ -58,7 +58,7 @@ public:
     Vec3 normal () const {return _normal;}
     void display (std::ostream& os) const;
     Collide collide (const Ray &ray) const;
-    float calc (const Vec3 & point) const;
+    double calc (const Vec3 & point) const;
     virtual Color color (const Vec3 &v = Vec3 ()) const;
 protected:
     void analyseContent (std::stringstream& content);
@@ -69,11 +69,11 @@ private:
     friend class Cobic;
     Plane () {}//仅yon用来构造cobic中的plane数组。
     Vec3 _center, _normal, _dx, _dy;
-    float _modelDx, _modelDy;
+    double _modelDx, _modelDy;
     //dx和dy只是为了指定纹理时使用，用来确定纹理的大小和方向
-    float _dot_center_normal;
+    double _dot_center_normal;
 
-    float _d;
+    double _d;
     bool _needBoudingBox;
 };
 
@@ -105,17 +105,17 @@ class Cobic: public Object
 {
 public:
     Cobic (std::stringstream& content, Condutor* condutor = nullptr);
-    Cobic (const Vec3& center, const Vec3& dx, const Vec3& dy, const Vec3& dz, float a, float b, float c, Material* material = nullptr, Condutor* condutor = nullptr, bool need = false);
+    Cobic (const Vec3& center, const Vec3& dx, const Vec3& dy, const Vec3& dz, double a, double b, double c, Material* material = nullptr, Condutor* condutor = nullptr, bool need = false);
     virtual ~Cobic ();
     Collide collide (const Ray& ray) const;
     virtual Color color (const Vec3 &v = Vec3 ()) const;
     void display (std::ostream& os) const;
-    float xh () const {return _x_h;}
-    float xl () const {return _x_l;}
-    float yh () const {return _y_h;}
-    float yl () const {return _y_l;}
-    float zh () const {return _z_h;}
-    float zl () const {return _z_l;}
+    double xh () const {return _x_h;}
+    double xl () const {return _x_l;}
+    double yh () const {return _y_h;}
+    double yl () const {return _y_l;}
+    double zh () const {return _z_h;}
+    double zl () const {return _z_l;}
 protected:
     void analyseContent (std::stringstream& content);
     bool check ();
@@ -123,9 +123,9 @@ protected:
     void preHandle ();
 private:
     Vec3 _center, _dx, _dy, _dz;
-    float _a, _b, _c;
-    float _a_half, _b_half, _c_half;
-    float _x_h, _x_l, _y_h, _y_l, _z_h, _z_l;
+    double _a, _b, _c;
+    double _a_half, _b_half, _c_half;
+    double _x_h, _x_l, _y_h, _y_l, _z_h, _z_l;
     Plane _sides[2][3];
     bool _needBoudingBox;
 };
